+++ deploy.sh
@@ -1,3 +1,6 @@
+git checkout master
+git merge develop
+
 ssh digitalocean <<'ENDSSH'
 cd /home/www/cookie-monster
 git pull