+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 ### 1.040.10 _(2014-01-xx)_
 
 * Added timer next to the Lucky Alert deficits to show time before the deficit is filled
-* Added Reindeer value in statistics
+* Fixed Reindeer value in statistics
 * Added statistics about whether it's better to buy pledge or make a covenant
 * Added a _grey_ color for upgrades, to differentiate items with low BCI/no BCI
 * Added support for multiple title modifiers (and Reindeer title modifier added with it)